Move matchMedia helper to Utils folder
