+++ tests/Frontend/support/e2e.js
@@ -20,7 +20,7 @@ import "./commands/adminRoomTypesCommands.js";
 import "./commands/adminSettingsCommands.js";
 import "./commands/adminRolesCommands.js";
 import "@cypress/code-coverage/support";
-import { patchMatchMedia } from "./utils/matchMediaHelper.js";
+import { patchMatchMedia } from "../../Utils/cypress/matchMediaHelper.js";
 
 Cypress.on("uncaught:exception", (err) => {
   // Check if error should be ignored
