+++ tests/System/support/e2e.js
@@ -15,7 +15,7 @@
 
 // Import commands.js using ES2015 syntax:
 import "./commands.js";
-import { patchMatchMedia } from "./utils/matchMediaHelper.js";
+import { patchMatchMedia } from "../../Utils/cypress/matchMediaHelper.js";
 
 Cypress.on("window:before:load", (win) => {
   // Set default value for dark mode to false, can be overridden in tests using Cypress.expose("darkMode", true/false)
